requirements.txt
pandas
numpy
scipy
requests
websockets
websocket-client